Uninstall dotenv changed the way of queries are implemented to more simplified and cleaner codes
