--- VirtuMart/.env
@@ -1,6 +1,0 @@
-PORT = 3000
-DB_HOST = 'localhost',
-DB_USER = 'Mart',
-DB_PW = 'Mart1234',
-DB_NAME = 'virtumartdb'
-DB_PORT = 3306
