+++ VirtuMart/package.json
@@ -13,7 +13,6 @@
   },
   "dependencies": {
     "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
     "express": "^4.18.3",
     "express-session": "^1.18.0",
     "mysql": "^2.18.1",
