+++ VirtuMart/server/server.js
@@ -8,7 +8,6 @@ import process from "process";
 
 import * as queries from "./queries.js";
 
-
 const PORT = process.env.PORT || 3000;
 const app = express();
